End of the day. Remains an annoying segfault. Cf below.
The segfault (actually, catched in an exception) is due to an
uninitialized edge_p pointer in NodeRelation.
Why, it's a mystery.

Other problem: the general behaviour is strange.
Either it's a question of tunning the constants in constants.h or there's
smthg wrong somewhere else.

Last thing: edges had a strange tendency to disappear suddently.
Check if it's still the case.